requirements.txt
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai==0.3.2
selenium==4.18.1
selenium-stealth==1.0.6
webdriver-manager==4.0.1
python-dotenv==1.0.1
pydantic==2.6.4
beautifulsoup4==4.12.3
streamlit==1.32.0
chromium-chromedriver 